removed the metricsService.ts file and it's reference
