+++ src/services/calculateMetrics.ts
@@ -1,56 +1,94 @@
 // src/services/calculateMetrics.ts
 
-export const calculateMetrics = (scrapedData: any) => {
-    // --- SEO Metrics ---
+import { IContent } from "../models/ContentModel";
+
+interface MetricResults {
+    seo: {
+        actualTitle: string | null;
+        titlePresent: boolean;
+        titleLength: number;
+        actualMetaDescription: string | null;
+        metaDescriptionPresent: boolean;
+        metaDescriptionLength: number;
+        headingsCount: number;
+        contentKeywords: string[];
+        seoFriendlyUrl: boolean;
+        faviconPresent: boolean;
+        faviconUrl: string | null;
+        robotsTxtAccessible: boolean;
+        inPageLinks: number;
+        languageDeclared: boolean;
+    };
+    security: {
+        httpsEnabled: boolean;
+        mixedContent: boolean;
+        serverSignatureHidden: boolean;
+        hstsEnabled: boolean;
+    };
+    performance: {
+        pageSizeKb: number;
+        httpRequests: number;
+        textCompressionEnabled: boolean;
+    };
+    miscellaneous: {
+        metaViewportPresent: boolean;
+        characterSet: string | null;
+        sitemapAccessible: boolean;
+        textToHtmlRatio: number;
+    };
+}
+
+export const calculateMetrics = (data: IContent): MetricResults => {
+    const url = data.url;
+    const htmlContent = data.htmlContent;
+
+    const title = data.metadata.title || null;
+    const metaDescription = data.metadata.description || null;
+
     const seoMetrics = {
-        actualTitle: scrapedData.metadata.title || null, // Include actual title in SEO section
-        title: scrapedData.metadata.title || "Missing Title", // Use actual scraped title
-        titleLength: calculateLength(scrapedData.metadata.title), // Length of the title
-        actualMetaDescription: scrapedData.metadata.description || null, // Include actual meta-description in SEO section
-        metaDescription: scrapedData.metadata.description || "Missing Meta Description", // Use actual scraped meta-description
-        metaDescriptionLength: calculateLength(scrapedData.metadata.description), // Length of the meta-description
-        headingsCount: countHeadings(scrapedData.htmlContent), // Number of headings
-        seoFriendlyUrl: isSeoFriendlyUrl(scrapedData.url), // Whether URL is SEO-friendly
-        faviconPresent: scrapedData.favicon ? true : false, // Whether a favicon is present
-        faviconUrl: scrapedData.favicon || null, // Add favicon URL
-        robotsTxtAccessible: isRobotsTxtAccessible(scrapedData.url), // Whether robots.txt is accessible
-        inPageLinks: countInPageLinks(scrapedData.htmlContent), // Count of in-page links
-        keywordsPresent: scrapedData.metadata.keywords || "No keywords found", // Keywords in metadata
+        actualTitle: title,
+        titlePresent: title !== null,
+        titleLength: calculateLength(title),
+        title: title || "Unknown Title", // Ensure title is never undefined
+        actualMetaDescription: metaDescription,
+        metaDescriptionPresent: metaDescription !== null,
+        metaDescriptionLength: calculateLength(metaDescription),
+        metaDescription: metaDescription || "No Meta Description", // Ensure metaDescription is never undefined
+        headingsCount: (htmlContent.match(/<h[1-6]>/g) || []).length,
+        contentKeywords: extractKeywords(htmlContent),
+        seoFriendlyUrl: isSeoFriendlyUrl(url),
+        faviconPresent: data.favicon !== null,
+        faviconUrl: data.favicon || null,
+        robotsTxtAccessible: isRobotsTxtAccessible(url),
+        inPageLinks: (htmlContent.match(/<a /g) || []).length,
+        languageDeclared: htmlContent.includes('<html lang='),
+        keywordsPresent: extractKeywords(htmlContent).length > 0 ? "Yes" : "No", // Ensure keywordsPresent is always a string
     };
 
-    // --- Security Metrics ---
-    const securityMetrics = {
-        https: scrapedData.url.startsWith("https"), // HTTPS enabled check
-        mixedContent: hasMixedContent(scrapedData.htmlContent), // Mixed content check
-        serverSignatureHidden: true, // Placeholder for server signature validation
-        hstsEnabled: true, // Placeholder for HSTS validation
-    };
-
-    // --- Performance Metrics ---
-    const performanceMetrics = {
-        pageSizeKb: calculatePageSizeKb(scrapedData.htmlContent), // Page size in KB
-        textCompressionEnabled: scrapedData.htmlContent.includes("Content-Encoding: gzip"), // Placeholder for text compression
-        httpRequests: countHttpRequests(scrapedData.htmlContent), // Count of HTTP requests
-    };
-
-    // --- Miscellaneous Metrics ---
-    const miscellaneousMetrics = {
-        sitemap: isSitemapAccessible(scrapedData.url), // Sitemap accessibility check
-        metaViewportPresent: scrapedData.htmlContent.includes('<meta name="viewport"'), // Meta viewport check
-        characterSet: extractCharacterSet(scrapedData.htmlContent), // Character set extraction
-        textToHtmlRatio: calculateTextToHtmlRatio(scrapedData.htmlContent, scrapedData.textContent), // Text-to-HTML ratio
-    };
-
-    // Return the calculated metrics, including raw title and meta description in the SEO section
     return {
         seo: seoMetrics,
-        security: securityMetrics,
-        performance: performanceMetrics,
-        miscellaneous: miscellaneousMetrics,
+        security: {
+            httpsEnabled: url.startsWith("https://"),
+            mixedContent: hasMixedContent(htmlContent),
+            serverSignatureHidden: true,
+            hstsEnabled: true,
+        },
+        performance: {
+            pageSizeKb: Math.ceil(htmlContent.length / 1024),
+            httpRequests: countHttpRequests(htmlContent),
+            textCompressionEnabled: htmlContent.includes("Content-Encoding: gzip"),
+        },
+        miscellaneous: {
+            metaViewportPresent: htmlContent.includes('<meta name="viewport"'),
+            characterSet: extractCharacterSet(htmlContent),
+            sitemapAccessible: isSitemapAccessible(url),
+            textToHtmlRatio: calculateTextToHtmlRatio(htmlContent, data.textContent),
+        },
     };
 };
 
-// --- Helper Functions (unchanged) ---
+
+// --- Helper Functions ---
 
 /**
  * Calculate the length of a given string.
@@ -62,10 +100,14 @@ const calculateLength = (text: string | null | undefined): number => {
 };
 
 /**
- * Count the number of headings (H1-H6) in the HTML content.
+ * Extract keywords from the meta tag in the HTML content.
  */
-const countHeadings = (htmlContent: string): number => {
-    return (htmlContent.match(/<h[1-6]>/g) || []).length;
+const extractKeywords = (htmlContent: string): string[] => {
+    const metaTag = htmlContent.match(/<meta name="keywords" content="([^"]+)"/i);
+    if (metaTag && metaTag[1]) {
+        return metaTag[1].split(",").map((keyword) => keyword.trim());
+    }
+    return [];
 };
 
 /**
@@ -84,16 +126,9 @@ const isRobotsTxtAccessible = (url: string): boolean => {
     try {
         const robotsTxtUrl = new URL("/robots.txt", url).href;
         return true; // Placeholder for actual HTTP request logic
-    } catch (error) {
+    } catch {
         return false;
     }
-};
-
-/**
- * Count the number of in-page links (anchor tags) in the HTML content.
- */
-const countInPageLinks = (htmlContent: string): number => {
-    return (htmlContent.match(/<a /g) || []).length;
 };
 
 /**
@@ -106,14 +141,7 @@ const hasMixedContent = (htmlContent: string): boolean => {
 };
 
 /**
- * Calculate the page size in kilobytes (KB).
- */
-const calculatePageSizeKb = (htmlContent: string): number => {
-    return Math.ceil(htmlContent.length / 1024); // Convert bytes to KB
-};
-
-/**
- * Count the number of HTTP requests (e.g., for scripts, images, links) in the HTML content.
+ * Count the number of HTTP requests in the HTML content.
  */
 const countHttpRequests = (htmlContent: string): number => {
     const links = (htmlContent.match(/<link /g) || []).length;
@@ -123,11 +151,11 @@ const countHttpRequests = (htmlContent: string): number => {
 };
 
 /**
- * Extract the character set from the HTML content (e.g., UTF-8).
+ * Extract the character set from the HTML content.
  */
 const extractCharacterSet = (htmlContent: string): string | null => {
-    const match = htmlContent.match(/<meta charset="([^"]+)"/);
-    return match ? match[1] : null;
+    const charsetMatch = htmlContent.match(/<meta charset="([^"]+)"/i);
+    return charsetMatch ? charsetMatch[1] : null;
 };
 
 /**
@@ -137,7 +165,7 @@ const isSitemapAccessible = (url: string): boolean => {
     try {
         const sitemapUrl = new URL("/sitemap.xml", url).href;
         return true; // Placeholder for actual HTTP request logic
-    } catch (error) {
+    } catch {
         return false;
     }
 };
